make discord helper more modular
some needed trimming, also makes all webhook funcs work in all channels
