+++ helpers/discord.js
@@ -1,53 +1,12 @@
-const config = require('../config.json');
 const { default: axios } = require('axios');
+const getWebhook = require('./discord/getWebhook');
+const getRoles = require('./discord/getRoles');
+const webhookColors = require('./discord/webhookColors');
 
 async function webhookPost(message, webhook) {
-    let url = 'https://discordapp.com/api/webhooks/';
-
-    switch (webhook) {
-        case 'osu':
-            url += `${config.standardWebhook.id}/${config.standardWebhook.token}`;
-            break;
-        case 'taiko':
-            url += `${config.taikoWebhook.id}/${config.taikoWebhook.token}`;
-            break;
-        case 'catch':
-            url += `${config.catchWebhook.id}/${config.catchWebhook.token}`;
-            break;
-        case 'mania':
-            url += `${config.maniaWebhook.id}/${config.maniaWebhook.token}`;
-            break;
-        case 'all':
-            url += `${config.allWebhook.id}/${config.allWebhook.token}`;
-            break;
-        case 'standardQualityAssurance':
-            url += `${config.standardQualityAssuranceWebhook.id}/${config.standardQualityAssuranceWebhook.token}`;
-            break;
-        case 'taikoCatchManiaQualityAssurance':
-            url += `${config.taikoCatchManiaQualityAssuranceWebhook.id}/${config.taikoCatchManiaQualityAssuranceWebhook.token}`;
-            break;
-        case 'beatmapReport':
-            url += `${config.beatmapReportWebhook.id}/${config.beatmapReportWebhook.token}`;
-            break;
-        case 'natUserReport':
-            url += `${config.natReportWebhook.id}/${config.natReportWebhook.token}`;
-            break;
-        case 'contentCase':
-            url += `${config.contentCasesWebhook.id}/${config.contentCasesWebhook.token}`;
-            break;
-        case 'internalContentCase':
-            url += `${config.internalContentCasesWebhook.id}/${config.internalContentCasesWebhook.token}`;
-            break;
-        case 'dev':
-            url += `${config.devWebhook.id}/${config.devWebhook.token}`;
-            break;
-        case 'announcement':
-            url += `${config.announcementWebhook.id}/${config.announcementWebhook.token}`;
-            break;
-        default:
-            return { error: 'no webhook specified' };
-    }
-
+    let url = getWebhook(webhook);
+    if (typeof url === 'object') return url;
+    
     try {
         await axios.post(url, {
             embeds: message,
@@ -58,27 +17,9 @@ async function webhookPost(message, webhook) {
 }
 
 async function highlightWebhookPost(message, webhook) {
-    let url = 'https://discordapp.com/api/webhooks/';
-    let role;
-
-    switch (webhook) {
-        case 'osuBeatmapReport':
-            url += `${config.beatmapReportWebhook.id}/${config.beatmapReportWebhook.token}`;
-            role = '<@&' + config.beatmapReportWebhook.osuRole + '>';
-            break;
-        case 'taikoBeatmapReport':
-            url += `${config.beatmapReportWebhook.id}/${config.beatmapReportWebhook.token}`;
-            role = '<@&' + config.beatmapReportWebhook.taikoRole + '>';
-            break;
-        case 'catchBeatmapReport':
-            url += `${config.beatmapReportWebhook.id}/${config.beatmapReportWebhook.token}`;
-            role = '<@&' + config.beatmapReportWebhook.catchRole + '>';
-            break;
-        case 'maniaBeatmapReport':
-            url += `${config.beatmapReportWebhook.id}/${config.beatmapReportWebhook.token}`;
-            role = '<@&' + config.beatmapReportWebhook.maniaRole + '>';
-            break;
-    }
+    const url = getWebhook(webhook);
+    const role = getRoles([webhook])
+    if (typeof url === 'object') return url;
 
     try {
         await axios.post(url, {
@@ -90,25 +31,8 @@ async function highlightWebhookPost(message, webhook) {
 }
 
 async function userHighlightWebhookPost(webhook, discordIds, text) {
-    let url = 'https://discordapp.com/api/webhooks/';
-
-    switch (webhook) {
-        case 'osu':
-            url += `${config.standardWebhook.id}/${config.standardWebhook.token}`;
-            break;
-        case 'taiko':
-            url += `${config.taikoWebhook.id}/${config.taikoWebhook.token}`;
-            break;
-        case 'catch':
-            url += `${config.catchWebhook.id}/${config.catchWebhook.token}`;
-            break;
-        case 'mania':
-            url += `${config.maniaWebhook.id}/${config.maniaWebhook.token}`;
-            break;
-        case 'all':
-            url += `${config.allWebhook.id}/${config.allWebhook.token}`;
-            break;
-    }
+    const url = getWebhook(webhook);
+    if (typeof url === 'object') return url;
 
     if (!text) text = '';
 
@@ -126,113 +50,11 @@ async function userHighlightWebhookPost(webhook, discordIds, text) {
 }
 
 async function roleHighlightWebhookPost(webhook, roles, text) {
-    let url = 'https://discordapp.com/api/webhooks/';
-    let content = '';
+    const url = getWebhook(webhook);
+    let content = getRoles(roles);
+    if (typeof url === 'object') return url;
 
-    switch (webhook) {
-        case 'natUserReport':
-            url += `${config.natReportWebhook.id}/${config.natReportWebhook.token}`;
-            break;
-        case 'contentCase':
-            url += `${config.contentCasesWebhook.id}/${config.contentCasesWebhook.token}`;
-            break;
-        case 'internalContentCase':
-            url += `${config.internalContentCasesWebhook.id}/${config.internalContentCasesWebhook.token}`;
-            break;
-        case 'announcement':
-            url += `${config.announcementWebhook.id}/${config.announcementWebhook.token}`;
-            break;
-        case 'osu':
-            url += `${config.standardWebhook.id}/${config.standardWebhook.token}`;
-            break;
-        case 'taiko':
-            url += `${config.taikoWebhook.id}/${config.taikoWebhook.token}`;
-            break;
-        case 'catch':
-            url += `${config.catchWebhook.id}/${config.catchWebhook.token}`;
-            break;
-        case 'mania':
-            url += `${config.maniaWebhook.id}/${config.maniaWebhook.token}`;
-            break;
-        case 'all':
-            url += `${config.allWebhook.id}/${config.allWebhook.token}`;
-            break;
-    }
-
-    for (const role of roles) {
-        switch (role) {
-            case 'gmt':
-                content += `<@&${config.announcementWebhook.gmtRole}> `;
-                break;
-            case 'contentReview':
-                content += `<@&${config.announcementWebhook.contentReviewRole}> `;
-                break;
-            case 'natInternal':
-                content += `<@&${config.announcementWebhook.natInternalRole}> `;
-                break;
-            case 'nat':
-                content += `<@&${config.announcementWebhook.natRole}> `;
-                break;
-            case 'natOsu':
-                content += `<@&${config.announcementWebhook.natOsu}> `;
-                break;
-            case 'natTaiko':
-                content += `<@&${config.announcementWebhook.natTaiko}> `;
-                break;
-            case 'natCatch':
-                content += `<@&${config.announcementWebhook.natCatch}> `;
-                break;
-            case 'natMania':
-                content += `<@&${config.announcementWebhook.natMania}> `;
-                break;
-            case 'natEvaluatorOsu':
-                content += `<@&${config.announcementWebhook.natEvaluatorOsu}> `;
-                break;
-            case 'natEvaluatorTaiko':
-                content += `<@&${config.announcementWebhook.natEvaluatorTaiko}> `;
-                break;
-            case 'natEvaluatorCatch':
-                content += `<@&${config.announcementWebhook.natEvaluatorCatch}> `;
-                break;
-            case 'natEvaluatorMania':
-                content += `<@&${config.announcementWebhook.natEvaluatorMania}> `;
-                break;
-            case 'bn':
-                content += `<@&${config.announcementWebhook.bnRole}> `;
-                break;
-            case 'bnOsu':
-                content += `<@&${config.announcementWebhook.bnOsu}> `;
-                break;
-            case 'bnTaiko':
-                content += `<@&${config.announcementWebhook.bnTaiko}> `;
-                break;
-            case 'bnCatch':
-                content += `<@&${config.announcementWebhook.bnCatch}> `;
-                break;
-            case 'bnMania':
-                content += `<@&${config.announcementWebhook.bnMania}> `;
-                break;
-            case 'probation':
-                content += `<@&${config.announcementWebhook.probationRole}> `;
-                break;
-            case 'probationOsu':
-                content += `<@&${config.announcementWebhook.probationOsu}> `;
-                break;
-            case 'probationTaiko':
-                content += `<@&${config.announcementWebhook.probationTaiko}> `;
-                break;
-            case 'probationCatch':
-                content += `<@&${config.announcementWebhook.probationCatch}> `;
-                break;
-            case 'probationMania':
-                content += `<@&${config.announcementWebhook.probationMania}> `;
-                break;
-            case 'groupMovers':
-                content += `<@&${config.announcementWebhook.groupMoversRole}> `;
-        }
-    }
-
-    if (text) content = `${content.trim()} ${text}`;
+    if (text) content += ` ${text}`;
 
     try {
         await axios.post(url, {
@@ -250,40 +72,6 @@ function defaultWebhookAuthor(session) {
         url: `https://osu.ppy.sh/users/${session.osuId}`,
     };
 }
-
-const webhookColors = {
-    lightRed: 16742771,       // active: moveToGroupDiscussion, almostOverdueNotifications
-    darkRed: 8787477,         // report, toggleIsSecurityChecked
-    red: 15607337,            // overdueNotifications, problemReport, lowActivity
-
-    lightOrange: 15639928,    // enableBnEvaluators, suggestionReport
-    darkOrange: 7092736,      // replaceMediator, pending badge
-    orange: 15169835,         // replaceEvaluator, switchBnEvaluator
-
-    lightYellow: 16777104,    // rejoinBnRequest
-    darkYellow: 7105536,      // concludeDiscussionVote
-    yellow: 16777022,         // submitMediationDiscussionVote
-
-    lightGreen: 8847214,      // submitEval, submitQA
-    darkGreen: 1921053,       // addToNat
-    green: 4380222,           // newBnApplication
-
-    lightBlue: 8643583,       // setConsensus, announcement
-    darkBlue: 1911891,        // setCooldown, overwriteEvaluationDate
-    blue: 6786559,            // setFeedback
-
-    lightPurple: 11173873,    // submitVetoMediation, toggleIsReviewed
-    darkPurple: 4263999,      // submitVeto
-    purple: 8536232,          // startVetoMediation, concludeVetoMediation
-
-    pink: 16728232,           // revealedCurrentBnEvalNotification, modified report, toggleHasTrialNat
-    lightPink: 16748236,      // isTrialNat, SEV, toggleIsBannedFromBn, editBadgeValue
-
-    white: 15724527,          // unarchive, addEvaluation, sendMessages
-    brown: 7554849,           // submitUserNote
-    gray: 8815494,            // passive: moveToGroupDiscussion, dev
-    black: 2564903,           // archive, deleteVeto
-};
 
 async function discussionWebhookPost(d, session) {
     const agree = d.mediations.filter(m => m.vote == 1);
